Refactor: API 코드 모듈화
